move commands back to kyma-folder
