+++ cmd/gendocs/gendocs.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd"
+	"github.com/kyma-project/cli/cmd/kyma/kyma"
 	"github.com/kyma-project/cli/internal/cli"
 	"github.com/spf13/cobra/doc"
 )
@@ -12,7 +12,7 @@ import (
 const DocsTargetDir = "./docs/gen-docs"
 
 func main() {
-	command := kymacmd.NewCmd(cli.NewOptions())
+	command := kyma.NewCmd(cli.NewOptions())
 	err := doc.GenMarkdownTree(command, DocsTargetDir)
 	if err != nil {
 		fmt.Println("unable to generate docs", err.Error())
