+++ cmd/kyma/install/cfg_windows.go
