+++ cmd/kyma/install/cmd_test.go
