+++ cmd/kyma/kyma/kyma.go
@@ -1,23 +1,22 @@
-package kymacmd
+package kyma
 
 import (
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/completion"
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/console"
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/install"
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/provision/azure"
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/provision/gardener"
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/provision/gcp"
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/provision/minikube"
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/test"
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/test/definitions"
-	del "github.com/kyma-project/cli/cmd/kyma/kymacmd/test/delete"
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/test/list"
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/test/logs"
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/test/run"
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/test/status"
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/version"
-
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/provision"
+	"github.com/kyma-project/cli/cmd/kyma/completion"
+	"github.com/kyma-project/cli/cmd/kyma/console"
+	"github.com/kyma-project/cli/cmd/kyma/install"
+	"github.com/kyma-project/cli/cmd/kyma/provision"
+	"github.com/kyma-project/cli/cmd/kyma/provision/azure"
+	"github.com/kyma-project/cli/cmd/kyma/provision/gardener"
+	"github.com/kyma-project/cli/cmd/kyma/provision/gcp"
+	"github.com/kyma-project/cli/cmd/kyma/provision/minikube"
+	"github.com/kyma-project/cli/cmd/kyma/test"
+	"github.com/kyma-project/cli/cmd/kyma/test/definitions"
+	del "github.com/kyma-project/cli/cmd/kyma/test/delete"
+	"github.com/kyma-project/cli/cmd/kyma/test/list"
+	"github.com/kyma-project/cli/cmd/kyma/test/logs"
+	"github.com/kyma-project/cli/cmd/kyma/test/run"
+	"github.com/kyma-project/cli/cmd/kyma/test/status"
+	"github.com/kyma-project/cli/cmd/kyma/version"
 	"github.com/kyma-project/cli/internal/cli"
 	"github.com/spf13/cobra"
 )
