+++ cmd/kyma/provision/azure/cmd_test.go
