+++ cmd/kyma/test/delete/cmd_test.go
