+++ cmd/kyma/test/logs/cmd.go
@@ -3,11 +3,12 @@ package logs
 import (
 	"bufio"
 	"fmt"
-	"github.com/kyma-project/cli/internal/logs"
 	"strings"
 
+	"github.com/kyma-project/cli/internal/logs"
+
 	oct "github.com/kyma-incubator/octopus/pkg/apis/testing/v1alpha1"
-	"github.com/kyma-project/cli/cmd/kyma/kymacmd/test"
+	"github.com/kyma-project/cli/cmd/kyma/test"
 	"github.com/kyma-project/cli/internal/cli"
 	"github.com/kyma-project/cli/internal/kube"
 	"github.com/pkg/errors"
