+++ cmd/kyma/test/status/cmd_test.go
